# Add spatial index for objects