--- a/irr/include/vector3d.h
+++ b/irr/include/vector3d.h
@@ -7,6 +7,7 @@
 #include "irrMath.h"
 
 #include <functional>
+#include <array>
 
 namespace irr
 {
@@ -32,6 +33,9 @@ public:
 	//! Constructor with the same value for all elements
 	explicit constexpr vector3d(T n) :
 			X(n), Y(n), Z(n) {}
+	//! Array - vector conversion
+	constexpr vector3d(const std::array<T, 3> &arr) :
+			X(arr[0]), Y(arr[1]), Z(arr[2]) {}
 
 	template <class U>
 	constexpr static vector3d<T> from(const vector3d<U> &other)
@@ -185,6 +189,10 @@ public:
 		Y = p.Y;
 		Z = p.Z;
 		return *this;
+	}
+
+	std::array<T, 3> toArray() const {
+		return {X, Y, Z};
 	}
 
 	//! Get length of the vector.