--- a/src/server/activeobjectmgr.cpp
+++ b/src/server/activeobjectmgr.cpp
@@ -116,13 +116,14 @@ void ActiveObjectMgr::invalidateActiveObjectObserverCaches()
 	}
 }
 
-void ActiveObjectMgr::updatePos(u16 id, const v3f &pos) {
+void ActiveObjectMgr::updateObjectPos(u16 id, v3f pos)
+{
 	// HACK only update if we already know the object
 	if (m_active_objects.get(id) != nullptr)
 		m_spatial_index.update(pos.toArray(), id);
 }
 
-void ActiveObjectMgr::getObjectsInsideRadius(const v3f &pos, float radius,
+void ActiveObjectMgr::getObjectsInsideRadius(v3f pos, float radius,
 		std::vector<ServerActiveObject *> &result,
 		std::function<bool(ServerActiveObject *obj)> include_obj_cb)
 {
@@ -153,7 +154,7 @@ void ActiveObjectMgr::getObjectsInArea(const aabb3f &box,
 }
 
 void ActiveObjectMgr::getAddedActiveObjectsAroundPos(
-		const v3f &player_pos, const std::string &player_name,
+		v3f player_pos, const std::string &player_name,
 		f32 radius, f32 player_radius,
 		const std::set<u16> &current_objects,
 		std::vector<u16> &added_objects)