--- a/src/server/activeobjectmgr.h
+++ b/src/server/activeobjectmgr.h
@@ -8,6 +8,7 @@
 #include <vector>
 #include "../activeobjectmgr.h"
 #include "serveractiveobject.h"
+#include "util/k_d_tree.h"
 
 namespace server
 {
@@ -25,6 +26,8 @@ public:
 
 	void invalidateActiveObjectObserverCaches();
 
+	void updatePos(const v3f &pos, u16 id);
+
 	void getObjectsInsideRadius(const v3f &pos, float radius,
 			std::vector<ServerActiveObject *> &result,
 			std::function<bool(ServerActiveObject *obj)> include_obj_cb);
@@ -36,5 +39,7 @@ public:
 			f32 radius, f32 player_radius,
 			const std::set<u16> &current_objects,
 			std::vector<u16> &added_objects);
+private:
+	DynamicKdTrees<3, f32, u16> m_spatial_index;
 };
 } // namespace server