--- a/src/server/activeobjectmgr.h
+++ b/src/server/activeobjectmgr.h
@@ -26,16 +26,16 @@ public:
 
 	void invalidateActiveObjectObserverCaches();
 
-	void updatePos(u16 id, const v3f &pos);
+	void updateObjectPos(u16 id, v3f pos);
 
-	void getObjectsInsideRadius(const v3f &pos, float radius,
+	void getObjectsInsideRadius(v3f pos, float radius,
 			std::vector<ServerActiveObject *> &result,
 			std::function<bool(ServerActiveObject *obj)> include_obj_cb);
 	void getObjectsInArea(const aabb3f &box,
 			std::vector<ServerActiveObject *> &result,
 			std::function<bool(ServerActiveObject *obj)> include_obj_cb);
 	void getAddedActiveObjectsAroundPos(
-			const v3f &player_pos, const std::string &player_name,
+			v3f player_pos, const std::string &player_name,
 			f32 radius, f32 player_radius,
 			const std::set<u16> &current_objects,
 			std::vector<u16> &added_objects);