--- a/src/server/serveractiveobject.cpp
+++ b/src/server/serveractiveobject.cpp
@@ -14,6 +14,12 @@ ServerActiveObject::ServerActiveObject(ServerEnvironment *env, v3f pos):
 {
 }
 
+void ServerActiveObject::setBasePosition(v3f pos) {
+	m_base_position = pos;
+	if (m_env) // HACK this doesn't feel right; *when* is m_env null?
+		ServerEnvironment_updatePos(m_env, pos, getId());
+}
+
 float ServerActiveObject::getMinimumSavedMovement()
 {
 	return 2.0*BS;