--- a/src/server/serveractiveobject.cpp
+++ b/src/server/serveractiveobject.cpp
@@ -18,8 +18,11 @@ ServerActiveObject::ServerActiveObject(ServerEnvironment *env, v3f pos):
 void ServerActiveObject::setBasePosition(v3f pos) {
 	bool changed = m_base_position != pos;
 	m_base_position = pos;
-	if (changed && getEnv()) // HACK *when* is getEnv() null?
+	if (changed && getEnv()) {
+		// getEnv() should never be null if the object is in an environment.
+		// It may however be null e.g. in tests or database migrations.
 		getEnv()->updateObjectPos(getId(), pos);
+	}
 }
 
 float ServerActiveObject::getMinimumSavedMovement()