--- a/src/serverenvironment.cpp
+++ b/src/serverenvironment.cpp
@@ -2061,8 +2061,3 @@ bool ServerEnvironment::migrateAuthDatabase(
 	}
 	return true;
 }
-
-// HACK
-void ServerEnvironment_updatePos(ServerEnvironment *senv, const v3f &pos, u16 id) {
-	senv->updatePos(pos, id);
-}