--- a/src/serverenvironment.h
+++ b/src/serverenvironment.h
@@ -220,8 +220,8 @@ public:
 	// Find the daylight value at pos with a Depth First Search
 	u8 findSunlight(v3s16 pos) const;
 
-	void updatePos(const v3f &pos, u16 id) {
-		return m_ao_manager.updatePos(pos, id);
+	void updatePos(u16 id, const v3f &pos) {
+		return m_ao_manager.updatePos(id, pos);
 	}
 
 	// Find all active objects inside a radius around a point
@@ -416,6 +416,3 @@ private:
 	std::unique_ptr<ServerActiveObject> createSAO(ActiveObjectType type, v3f pos,
 			const std::string &data);
 };
-
-// HACK
-void ServerEnvironment_updatePos(ServerEnvironment *senv, const v3f &pos, u16 id);