@sfan5 2024-05-09
I think this will need comprehensive tests for how the spatial index and object lifecycle interact.