@ExeVirus 2024-06-03
Mostly we need solid, repeatable, verifiable, and somewhat diverse test
methods, to prove to ourselves these implementations are flawless, and to
have a very good benchmark to say which we choose.


There are small scale tests we can set up easily, but the most important is
one that is pseudorandom, that adds, removes, and moves entities, starting
with 100, up to about 2-3K, and back down.

We would do both in radius and in box checks during this time (say every 20
frames or something), and I'd want to see identical results between master
and these branches before saying merge. We haven't really tested all the
cases of entities being added and removed during iteration in real time and
need to be confident in the solution.

I think it would take maybe 3-5 hours to get those tests written, and I
would have them done probably within a week or two.

After we verify they are safe implementations, it should be very easy to
compare performance and consider code maintenance at that point to decide
the better solution.

Really don't want to over engineer, or bike shed, but hey as long as a core
dev doesn't have to do it... Shrug



On Mon, Jun 3, 2024, 12:36 PM lhofhansl ***@***.***> wrote:

> What's needed to make progress on this one (or #14643
> <https://github.com/minetest/minetest/pull/14643>)?
> I tried both out on a busy world and it all seems to work. (In my scenario
> I did not see any performance improvements - many object in a single blocks
> - but all continued to work fine)
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/minetest/minetest/pull/14631#issuecomment-2145660616>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AKT7N2WUEVIM7BTOEKYFALDZFSLRFAVCNFSM6AAAAABHPG5SFWVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDCNBVGY3DANRRGY>
> .
> You are receiving this because you commented.Message ID:
> ***@***.***>
>